+++ st2common/st2common/util/jinja.py
@@ -60,7 +60,7 @@ def get_filters():
     from st2common.jinja.filters import time
     from st2common.jinja.filters import version
     from st2common.jinja.filters import json_escape
-    from st2common.jinja.filters import jmespath_query
+    from st2common.jinja.filters import jsonpath_query
 
     # IMPORTANT NOTE - these filters were recently duplicated in st2mistral so that
     # they are also available in Mistral workflows. Please ensure any additions you
@@ -93,7 +93,7 @@ def get_filters():
         'use_none': use_none,
 
         'json_escape': json_escape.json_escape,
-        'jmespath_query': jmespath_query.jmespath_query
+        'jsonpath_query': jsonpath_query.jsonpath_query
     }