+++ st2common/tests/unit/test_jinja_render_jsonpath_query_filters.py
@@ -19,9 +19,9 @@ import unittest2
 from st2common.util import jinja as jinja_utils
 
 
-class JinjaUtilsJmespathQueryTestCase(unittest2.TestCase):
+class JinjaUtilsJsonpathQueryTestCase(unittest2.TestCase):
 
-    def test_jmespath_query_static(self):
+    def test_jsonpath_query_static(self):
         env = jinja_utils.get_jinja_environment()
         obj = {'people': [{'first': 'James', 'last': 'd'},
                           {'first': 'Jacob', 'last': 'e'},
@@ -29,13 +29,13 @@ class JinjaUtilsJmespathQueryTestCase(unittest2.TestCase):
                           {'missing': 'different'}],
                'foo': {'bar': 'baz'}}
 
-        template = '{{ obj | jmespath_query("people[*].first") }}'
+        template = '{{ obj | jsonpath_query("people[*].first") }}'
         actual_str = env.from_string(template).render({'obj': obj})
         actual = eval(actual_str)
         expected = ['James', 'Jacob', 'Jayden']
         self.assertEqual(actual, expected)
 
-    def test_jmespath_query_dynamic(self):
+    def test_jsonpath_query_dynamic(self):
         env = jinja_utils.get_jinja_environment()
         obj = {'people': [{'first': 'James', 'last': 'd'},
                           {'first': 'Jacob', 'last': 'e'},
@@ -44,9 +44,25 @@ class JinjaUtilsJmespathQueryTestCase(unittest2.TestCase):
                'foo': {'bar': 'baz'}}
         query = "people[*].last"
 
-        template = '{{ obj | jmespath_query(query) }}'
+        template = '{{ obj | jsonpath_query(query) }}'
         actual_str = env.from_string(template).render({'obj': obj,
                                                        'query': query})
         actual = eval(actual_str)
         expected = ['d', 'e', 'f']
+        self.assertEqual(actual, expected)
+
+    def test_jsonpath_query_no_results(self):
+        env = jinja_utils.get_jinja_environment()
+        obj = {'people': [{'first': 'James', 'last': 'd'},
+                          {'first': 'Jacob', 'last': 'e'},
+                          {'first': 'Jayden', 'last': 'f'},
+                          {'missing': 'different'}],
+               'foo': {'bar': 'baz'}}
+        query = "query_returns_no_results"
+
+        template = '{{ obj | jsonpath_query(query) }}'
+        actual_str = env.from_string(template).render({'obj': obj,
+                                                       'query': query})
+        actual = eval(actual_str)
+        expected = None
         self.assertEqual(actual, expected)